requirements.txt
numpy==2.4.2
opencv_python==4.12.0.88
Pillow==12.1.0
streamlit==1.53.1
tensorflow==2.20.0
